Removing pointless helpers utils
