--- src/shared/utils/helpers/capitalize-first-letter.ts
@@ -1,3 +1,0 @@
-export default function capitalizeFirstLetter(str: string): string {
-  return str.charAt(0).toUpperCase() + str.slice(1);
-}
