--- src/shared/utils/helpers/get-bool-from-string.ts
@@ -1,5 +1,0 @@
-export default function getBoolFromString(
-  boolStr?: string,
-): boolean | undefined {
-  return boolStr ? boolStr.toLowerCase() === "true" : undefined;
-}
