--- src/shared/utils/helpers/get-id-from-string.ts
@@ -1,3 +1,0 @@
-export default function getIdFromString(id?: string): number | undefined {
-  return id && id !== "0" && !Number.isNaN(Number(id)) ? Number(id) : undefined;
-}
