--- src/shared/utils/helpers/get-page-from-string.ts
@@ -1,3 +1,0 @@
-export default function getPageFromString(page?: string): number {
-  return page && !Number.isNaN(Number(page)) ? Number(page) : 1;
-}
