--- src/shared/utils/helpers/get-random-char-from-alphabet.ts
@@ -1,3 +1,0 @@
-export default function getRandomCharFromAlphabet(alphabet: string): string {
-  return alphabet.charAt(Math.floor(Math.random() * alphabet.length));
-}
