--- src/shared/utils/helpers/get-random-from-list.ts
@@ -1,5 +1,0 @@
-export default function getRandomFromList<T>(list: T[]): T | undefined {
-  return list.length === 0
-    ? undefined
-    : list.at(Math.floor(Math.random() * list.length));
-}
