--- src/shared/utils/helpers/group-by.ts
@@ -1,11 +1,0 @@
-export const groupBy = <T>(
-  array: T[],
-  predicate: (value: T, index: number, array: T[]) => string,
-) =>
-  array.reduce(
-    (acc, value, index, array) => {
-      (acc[predicate(value, index, array)] ||= []).push(value);
-      return acc;
-    },
-    {} as { [key: string]: T[] },
-  );
