--- src/shared/utils/helpers/hostname.ts
@@ -1,4 +1,0 @@
-export default function hostname(url: string): string {
-  const cUrl = new URL(url);
-  return cUrl.port ? `${cUrl.hostname}:${cUrl.port}` : `${cUrl.hostname}`;
-}
