--- src/shared/utils/helpers/hsl.ts
@@ -1,3 +1,0 @@
-export default function hsl(num: number) {
-  return `hsla(${num}, 35%, 50%, 0.5)`;
-}
