--- src/shared/utils/helpers/random-str.ts
@@ -1,19 +1,0 @@
-import { getRandomCharFromAlphabet } from "@utils/helpers";
-
-const DEFAULT_ALPHABET =
-  "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-
-export default function randomStr(
-  idDesiredLength = 20,
-  alphabet = DEFAULT_ALPHABET,
-): string {
-  /**
-   * Create n-long array and map it to random chars from given alphabet.
-   * Then join individual chars as string
-   */
-  return Array.from({ length: idDesiredLength })
-    .map(() => {
-      return getRandomCharFromAlphabet(alphabet);
-    })
-    .join("");
-}
