--- src/shared/utils/helpers/resources-settled.ts
@@ -1,5 +1,0 @@
-import { RequestState } from "../../services/HttpService";
-
-export default function resourcesSettled(resources: RequestState<any>[]) {
-  return resources.every(r => r.state === "success" || r.state === "failed");
-}
