--- src/shared/utils/helpers/sleep.ts
@@ -1,3 +1,0 @@
-export default function sleep(millis: number): Promise<void> {
-  return new Promise(resolve => setTimeout(resolve, millis));
-}
