--- src/shared/utils/helpers/valid-email.ts
@@ -1,5 +1,0 @@
-export default function validEmail(email: string) {
-  const re =
-    /^(([^\s"(),.:;<>@[\\\]]+(\.[^\s"(),.:;<>@[\\\]]+)*)|(".+"))@((\[(?:\d{1,3}\.){3}\d{1,3}])|(([\dA-Za-z\-]+\.)+[A-Za-z]{2,}))$/;
-  return re.test(String(email).toLowerCase());
-}
