--- src/shared/utils/helpers/valid-url.ts
@@ -1,8 +1,0 @@
-export default function validURL(str: string) {
-  try {
-    new URL(str);
-    return true;
-  } catch {
-    return false;
-  }
-}
